A number of changes to fixup metadata
This generator has stricter rules as far as referenced/implemented/base
classes/interfaces are concerned - all of them must be known, that is in one way
or another defined in either the API description file or the fixup files.

This commit moves some generic type declarations from the `name` attribute to
the `name-generic-aware`. This is more consistent with what we have in most
other types as well as makes the name explicitly generic or non-generic. When XA
supports Java generic types the generator will be able to use the
`name-generic-aware` attribute to generate the correct class.

References to generic types are also moved to various `*-generic-aware`
attributes.

Additionally, more references to `Java.Nio.Channels.{Network,Multicast}Channel`
are removed.